+++ gradle.properties
@@ -44,13 +44,13 @@ accessTransformersFile = visualprospecting_at.cfg
 # Provides setup for Mixins if enabled. If you don't know what mixins are: Keep it disabled!
 usesMixins = true
 # Specify the location of your implementation of IMixinConfigPlugin. Leave it empty otherwise.
-mixinPlugin = mixinplugin.MixinPlugin
+mixinPlugin =
 # Specify the package that contains all of your Mixins. You may only place Mixins in this package or the build will fail!
 mixinsPackage = mixins
 # Specify the core mod entry class if you use a core mod. This class must implement IFMLLoadingPlugin!
 # This parameter is for legacy compatibility only
 # Example value: coreModClass = asm.FMLPlugin + modGroup = com.myname.mymodid -> com.myname.mymodid.asm.FMLPlugin
-coreModClass =
+coreModClass = mixinplugin.VisualProspectingEarlyMixinLoader
 # If your project is only a consolidation of mixins or a core mod and does NOT contain a 'normal' mod ( = some class
 # that is annotated with @Mod) you want this to be true. When in doubt: leave it on false!
 containsMixinsAndOrCoreModOnly = false
