+++ src/main/java/com/sinthoras/visualprospecting/mixinplugin/TargetedMod.java
@@ -1,57 +1,35 @@
 package com.sinthoras.visualprospecting.mixinplugin;
 
-import java.nio.file.Path;
-
-import com.google.common.io.Files;
-
 public enum TargetedMod {
 
-    //
-    // IMPORTANT: Do not make any references to any mod from this file. This file is loaded quite early on and if
-    // you refer to other mods you load them as well. The consequence is: You can't inject any previously loaded
-    // classes!
-    // Exception: Tags.java, as long as it is used for Strings only!
-    //
+    BARTWORKS("BartWorks", "com.github.bartimaeusnek.bartworks.ASM.BWCorePlugin", "bartworks"),
+    GALACTICGREG("Galactic Greg", null, "galacticgreg"),
+    GT5U("GregTech5u", null, "gregtech"), // Also matches GT6.
+    JOURNEYMAP("JourneyMap", null, "journeymap"),
+    TCNODETRACKER("TC Node Tracker", null, "tcnodetracker"),
+    VANILLA("Minecraft", null),
+    XAEROMINIMAP("Xaero's Minimap", "xaero.common.core.XaeroMinimapPlugin", "XaeroMinimap"),
+    XAEROWORLDMAP("Xaero's World Map", "xaero.map.core.XaeroWorldMapPlugin", "XaeroWorldMap");
 
-    // Replace with your injected mods here, but always keep VANILLA:
-    VANILLA("Minecraft", "unused", true),
-    GREGTECH("GregTech", "gregtech", true),
-    JOURNEYMAP("JourneyMap", "journeymap-1.7.10", true),
-    XAEROWORLDMAP("Xaero's World Map", "XaerosWorldMap", true),
-    XAEROMINIMAP("Xaero's Minimap", "Xaeros_Minimap", true),
-    TCNODETRACKER("TCNodeTracker", "tcnodetracker-1.7.10", true),
-    BARTWORKS("Bartworks", "bartworks", false),
-    GALACTICGREG("GalacticGreg", "GalacticGreg", false);
-
+    /** The "name" in the @Mod annotation */
     public final String modName;
-    public final String jarNamePrefixLowercase;
-    // Optional dependencies can be omitted in development. Especially skipping GT5U will drastically speed up your game
-    // start!
-    public final boolean loadInDevelopment;
+    /** Class that implements the IFMLLoadingPlugin interface */
+    public final String coreModClass;
+    /** The "modid" in the @Mod annotation */
+    public final String modId;
 
-    TargetedMod(String modName, String jarNamePrefix, boolean loadInDevelopment) {
-        this.modName = modName;
-        this.jarNamePrefixLowercase = jarNamePrefix.toLowerCase();
-        this.loadInDevelopment = loadInDevelopment;
+    TargetedMod(String modName, String coreModClass) {
+        this(modName, coreModClass, null);
     }
 
-    public boolean isMatchingJar(Path path) {
-        final String pathString = path.toString();
-        final String nameLowerCase = Files.getNameWithoutExtension(pathString).toLowerCase();
-        final String fileExtension = Files.getFileExtension(pathString);
-
-        return nameLowerCase.startsWith(jarNamePrefixLowercase)
-                && ("jar".equals(fileExtension) || "litemod".equals(fileExtension));
+    TargetedMod(String modName, String coreModClass, String modId) {
+        this.modName = modName;
+        this.coreModClass = coreModClass;
+        this.modId = modId;
     }
 
     @Override
     public String toString() {
-        return "TargetedMod{" + "modName='"
-                + modName
-                + '\''
-                + ", jarNamePrefixLowercase='"
-                + jarNamePrefixLowercase
-                + '\''
-                + '}';
+        return "TargetedMod{modName='" + modName + "', coreModClass='" + coreModClass + "', modId='" + modId + "'}";
     }
 }
