+++ src/main/java/com/sinthoras/visualprospecting/mixinplugin/VisualProspectingEarlyMixinLoader.java
@@ -1,0 +1,51 @@
+package com.sinthoras.visualprospecting.mixinplugin;
+
+import java.util.*;
+
+import com.gtnewhorizon.gtnhmixins.IEarlyMixinLoader;
+import cpw.mods.fml.relauncher.IFMLLoadingPlugin;
+
+@IFMLLoadingPlugin.MCVersion("1.7.10")
+@IFMLLoadingPlugin.TransformerExclusions("com.sinthoras.visualprospecting.mixinplugin")
+public class VisualProspectingEarlyMixinLoader implements IFMLLoadingPlugin, IEarlyMixinLoader {
+
+    @Override
+    public String getMixinConfig() {
+        return "mixins.visualprospecting.early.json";
+    }
+
+    @Override
+    public List<String> getMixins(Set<String> loadedCoreMods) {
+        final List<String> mixins = new ArrayList<>();
+        for (Mixins mixin : Mixins.values()) {
+            if (mixin.phase == Mixins.Phase.EARLY && mixin.shouldLoad(loadedCoreMods, Collections.emptySet())) {
+                mixins.add(mixin.mixinClass);
+            }
+        }
+        return mixins;
+    }
+
+    @Override
+    public String[] getASMTransformerClass() {
+        return null;
+    }
+
+    @Override
+    public String getModContainerClass() {
+        return null;
+    }
+
+    @Override
+    public String getSetupClass() {
+        return null;
+    }
+
+    @Override
+    public void injectData(Map<String, Object> data) {}
+
+    @Override
+    public String getAccessTransformerClass() {
+        return null;
+    }
+
+}
