--- src/main/java/com/sinthoras/visualprospecting/mixins/journeymap/DisplayVarsAccessor.java
@@ -1,24 +1,0 @@
-package com.sinthoras.visualprospecting.mixins.journeymap;
-
-import journeymap.client.ui.minimap.DisplayVars;
-import journeymap.client.ui.minimap.Shape;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
-
-@Mixin(DisplayVars.class)
-public interface DisplayVarsAccessor {
-
-    @Accessor(remap = false)
-    float getDrawScale();
-
-    @Accessor(remap = false)
-    double getFontScale();
-
-    @Accessor(remap = false)
-    Shape getShape();
-
-    @Accessor(remap = false)
-    int getMinimapWidth();
-
-}
