+++ src/main/resources/mixins.visualprospecting.late.json
@@ -1,0 +1,8 @@
+{
+  "required": true,
+  "minVersion": "0.7.11",
+  "package": "com.sinthoras.visualprospecting.mixins.late",
+  "refmap": "mixins.visualprospecting.refmap.json",
+  "target": "@env(DEFAULT)",
+  "compatibilityLevel": "JAVA_8"
+}
